Reorganize code - move tests related to hyperkit to driver_install_or_update_darwin_test.go
