+++ test/integration/driver_install_or_update_darwin_test.go
@@ -9,9 +9,79 @@ import (
 	"testing"
 
 	"github.com/Azure/azure-sdk-for-go/tools/apidiff/ioext"
+	"github.com/blang/semver"
 
+	"k8s.io/minikube/pkg/minikube/driver"
 	"k8s.io/minikube/pkg/version"
 )
+
+func TestHyperKitDriverInstallOrUpdate(t *testing.T) {
+	MaybeParallel(t)
+
+	tests := []struct {
+		name string
+		path string
+	}{
+		{name: "driver-without-version-support", path: filepath.Join(*testdataDir, "hyperkit-driver-without-version")},
+		{name: "driver-with-older-version", path: filepath.Join(*testdataDir, "hyperkit-driver-without-version")},
+	}
+
+	originalPath := os.Getenv("PATH")
+	defer os.Setenv("PATH", originalPath)
+
+	for _, tc := range tests {
+		dir, err := ioutil.TempDir("", tc.name)
+		if err != nil {
+			t.Fatalf("Expected to create tempdir. test: %s, got: %v", tc.name, err)
+		}
+		defer func() {
+			err := os.RemoveAll(dir)
+			if err != nil {
+				t.Errorf("Failed to remove dir %q: %v", dir, err)
+			}
+		}()
+
+		pwd, err := os.Getwd()
+		if err != nil {
+			t.Fatalf("Error not expected when getting working directory. test: %s, got: %v", tc.name, err)
+		}
+
+		path := filepath.Join(pwd, tc.path)
+
+		_, err = os.Stat(filepath.Join(path, "docker-machine-driver-hyperkit"))
+		if err != nil {
+			t.Fatalf("Expected driver to exist. test: %s, got: %v", tc.name, err)
+		}
+
+		// change permission to allow driver to be executable
+		err = os.Chmod(filepath.Join(path, "docker-machine-driver-hyperkit"), 0700)
+		if err != nil {
+			t.Fatalf("Expected not expected when changing driver permission. test: %s, got: %v", tc.name, err)
+		}
+
+		os.Setenv("PATH", fmt.Sprintf("%s:%s", path, originalPath))
+
+		// NOTE: This should be a real version, as it impacts the downloaded URL
+		newerVersion, err := semver.Make("1.3.0")
+		if err != nil {
+			t.Fatalf("Expected new semver. test: %v, got: %v", tc.name, err)
+		}
+
+		if sudoNeedsPassword() {
+			t.Skipf("password required to execute 'sudo', skipping remaining test")
+		}
+
+		err = driver.InstallOrUpdate("hyperkit", dir, newerVersion, false, true)
+		if err != nil {
+			t.Fatalf("Failed to update driver to %v. test: %s, got: %v", newerVersion, tc.name, err)
+		}
+
+		_, err = os.Stat(filepath.Join(dir, "docker-machine-driver-hyperkit"))
+		if err != nil {
+			t.Fatalf("Expected driver to be download. test: %s, got: %v", tc.name, err)
+		}
+	}
+}
 
 func TestHyperkitDriverSkipUpgrade(t *testing.T) {
 	MaybeParallel(t)
@@ -72,6 +142,11 @@ func TestHyperkitDriverSkipUpgrade(t *testing.T) {
 			}
 		})
 	}
+}
+
+func sudoNeedsPassword() bool {
+	err := exec.Command("sudo", "-n", "ls").Run()
+	return err != nil
 }
 
 func driverVersion(path string) (string, error) {
